feat: added historicalPrices get method. Still needs correct implementation everywhere bugs: pages load 2 times and therefore httpcontext is null

+++ code/WebApp/WebApp/Models/Entities/WalletEntry.cs
@@ -5,8 +5,8 @@ namespace WebApp.Models.Entities
     public class WalletEntry
     {
         public string Name { get; set; }
-        public decimal? Amount { get; set; }
-        public decimal? AveragePrice { get; set; }
-        public decimal? CurrentPrice { get; set; }
+        public double? Amount { get; set; }
+        public double? AveragePrice { get; set; }
+        public double? CurrentPrice { get; set; }
     }
 }
